services/switchbot/requirements.txt
paho-mqtt>=2.0.0
aiohttp>=3.9.0
pyyaml>=6.0
